refactor: changed boss health to double and fixed health reduction

+++ Space-Invaders/src/Boss.java
@@ -1,13 +1,13 @@
 import java.awt.Image;
 
 public class Boss extends Block {
-    private int health;
+    private double health;
 
     public Boss(int x, int y, int width, int height, Image image) {
         super(x, y, width, height, image);
         this.health = 100;
     }
 
-    public int getHealth() { return health; }
-    public void setHealth(int health) { this.health = health; }
+    public double getHealth() { return health; }
+    public void setHealth(double health) { this.health = health; }
 }
